Routing fix, minor improvements
