+++ explorer-frontend/src/components/AppFooter.vue
@@ -55,7 +55,5 @@
 </template>
 
 <script setup lang="ts">
-import { useI18n } from "vue-i18n";
-
 const { t } = useI18n();
 </script>
