+++ explorer-frontend/src/components/SyncNotice.vue
@@ -20,7 +20,6 @@
 
 <script setup lang="ts">
 import { useBackgroundInfo, useBlockchainInfo } from "@/composables/States";
-import { useI18n } from "vue-i18n";
 
 const { t } = useI18n();
 const backgroundInfoDataState = useBackgroundInfo();
