+++ explorer-frontend/src/components/home/RecentBlocks.vue
@@ -14,7 +14,7 @@ const config = useRuntimeConfig();
 const latestBlock = useLatestBlockInfo();
 
 const recentBlocksInfo = await useFetch<Array<SimplifiedBlock>>(
-  `${getApiPath()}/blocks?offset=0&count=${config.RECENT_BLOCKS_COUNT}&sort=1`
+  `${getApiPath()}/blocks?offset=0&count=${config.public.recentBlocksCount}&sort=1`
 );
 const recentBlocks = reactive<Array<SimplifiedBlock>>(
   recentBlocksInfo.data.value ?? []
