+++ explorer-frontend/src/composables/Localization.ts
@@ -1,5 +1,3 @@
-import { useI18n } from "vue-i18n";
-
 export const useLocalization = () => {
 
     const headers = useRequestHeaders(["accept-language"]);
