+++ explorer-frontend/src/pages/index.vue
@@ -44,7 +44,7 @@ const meta = computed(() => {
       },
       {
         name: "og:url",
-        content: `${config.BASE_URL}/`,
+        content: `${config.public.baseUrl}/`,
       },
     ],
   };
